requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
